requirements.txt
pillow
mysql-connector-python
matplotlib
